PDFGeneration, Confirmation and Review update
